% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulationStudy.R
\name{runSimulationStudy}
\alias{runSimulationStudy}
\title{Runs Simulation Study}
\usage{
runSimulationStudy(simulationProfile, simulationSetup, cohortMethodData,
  simulationRuns = 10, trueEffectSize = NA, outcomePrevalence = NA,
  hdpsFeatures, stratify = FALSE, discrete = FALSE,
  ignoreCensoring = FALSE, threads = 10, fudge = 0.001, prior = NULL)
}
\arguments{
\item{simulationProfile}{simulationProfile object created by createCMDSimulationProfile function}

\item{simulationRuns}{Number of simulations to run (1 simulation = reroll outcomes)}

\item{trueEffectSize}{True effect size for exposure to simulate. If set to NULL keeps observed effect size from simulation profile}

\item{outcomePrevalence}{Outcome prevalence to simulate; adjusts outcome baseline survival function to achieve. If null keeps observed}

\item{hdpsFeatures}{TRUE = using HDPS features; FALSE = using FeatureExtraction features}

\item{ignoreCensoring}{Ignore censoring altogether; sets censoring process baseline survival function to 1}

\item{threads}{Number of parallel threads to use}

\item{studyPop}{Study population created by createStudyPopulation function}

\item{confoundingScheme}{Type of unmeasured confounding to use for PS (0 = none; 1 = demographics; 2 = random proportion; 3 = demographics and random proportion)}

\item{confoundingProportion}{Proportion of covariates to hide from propensity score as unmeasured confounding}

\item{ignoreCensoringCovariates}{Ignore covariates effects on censoring process; only uses baseline function}
}
\value{
Returns the following: \describe{
\item{trueOutcomeModel}{the non-exposure oefficients used in the outcome generating model}
\item{trueEffectSize}{coefficient for exposure effect}
\item{estimatesLasso}{logRr, bound, and sd for LASSO estimate for each simulation}
\item{estimatesExpHdps}{logRr, bound, and sd for exposure based hdps}
\item{estimatesBiasHdps}{logRr, bound, and sd for bias based hdps}
\item{aucLasso}{auc for LASSO propensity score}
\item{aucExpHdps}{auc for exposure based hdps propensity score}
\item{aucBiasHdps}{auc for bias based hdps propensity score}
\item{psLasso}{propensity scores for subjects in study population for LASSO}
\item{psExp}{propensity scores for subjects in study population for exposure based hdps}
\item{psBias}{propensity scores for subjects in study population for bias based hdps; propensity and preference scores are averaged over simulation runs}
\item{outcomePrevalence}{outcome prevalence of simulation}}
}
\description{
This function runs a simulation to compare LASSO, exposure based hdps, and bias based hdps as propensity score methods
}

