% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulationStudy.R
\name{runSimulationStudy}
\alias{runSimulationStudy}
\title{Runs Simulation Study}
\usage{
runSimulationStudy(cohortMethodData, confoundingScheme = 0,
  confoundingProportion = 0.3, n = 10, trueBeta = NULL,
  outcomePrevalence = NULL, crossValidate = TRUE, hdpsFeatures = FALSE,
  ignoreCensoring = FALSE, ignoreCensoringCovariates = TRUE)
}
\arguments{
\item{cohortMethodData}{cohortMethodData object}

\item{confoundingScheme}{Type of unmeasured confounding to use for PS (0 = none; 1 = demographics; 2 = random proportion; 3 = demographics and random proportion)}

\item{confoundingProportion}{Proportion of covariates to hide from propensity score as unmeasured confounding}

\item{n}{Number of simulations to run (1 simulation = reroll outcomes)}

\item{trueBeta}{True effect size for exposure to simulate}

\item{outcomePrevalence}{Outcome prevalence to simulate; adjusts outcome baseline survival function to achieve}

\item{crossValidate}{Can turn off cross validation when fitting outcome models in beginning and fitting LASSO propensity score}

\item{hdpsFeatures}{TRUE = using HDPS features; FALSE = using FeatureExtraction features}

\item{ignoreCensoring}{Ignore censoring altogether; sets censoring process baseline survival function to 1}

\item{ignoreCensoringCovariates}{Ignore covariates effects on censoring process; only uses baseline function}
}
\value{
Returns the following: \describe {
\item{trueOutcomeModel}{coefficients used for true outcome model}
\item{trueEffectSize}{coefficient for exposure effect}
\item{estimatesLasso}{logRr, bound, and sd for LASSO estimate for each simulation}
\item{estimatesExpHdps}{logRr, bound, and sd for exposure based hdps}
\item{estimatesBiasHdps}{logRr, bound, and sd for bias based hdps}
\item{aucLasso}{auc for LASSO propensity score}
\item{aucExpHdps}{auc for exposure based hdps propensity score}
\item{aucBiasHdps}{auc for bias based hdps propensity score}
\item{psLasso}{propensity scores for subjects in study population for LASSO}
\item{psExp}{propensity scores for subjects in study population for exposure based hdps}
\item{psBias}{propensity scores for subjects in study population for bias based hdps; propensity and preference scores are averaged over simulation runs}
\item{originalOutcomePrevalence}{original outcome prevalence before adjustment to desired outcome prevalence}}
}
\description{
This function runs a simulation to compare LASSO, exposure based hdps, and bias based hdps as propensity score methods
}

